# docs/swagger.yaml
basePath: /v1
definitions:
...
  games.BookingStatus:
    enum:
    - pending
    - requested
    - booked
    - rejected
    - cancelled
    type: string
    x-enum-varnames:
    - BookingPending
    - BookingRequested
    - BookingBooked
    - BookingRejected
    - BookingCancelled
...
  games.ShortlistedGameDetail:
    properties:
      admin_id:
        description: Foreign key to users table (game admin)
        type: integer
      booking_status:
        $ref: '#/definitions/games.BookingStatus'
      created_at:
        description: Timestamp when the game was created
        type: string
      end_time:
        description: Game end time
        type: string
      format:
        description: Game format (nullable)
        type: string
      game_level:
        description: Skill level (beginner, intermediate, advanced)
        type: string
      id:
        description: Primary key
        type: integer
      instruction:
        description: Game instructions (nullable)
        type: string
      match_full:
        description: Whether the game is full
        type: boolean
      max_players:
        description: Maximum number of players
        type: integer
      price:
        description: Price of the game (nullable)
        type: integer
      sport_type:
        description: Type of sport (e.g., futsal, basketball)
        type: string
      start_time:
        description: Game start time
        type: string
      status:
        description: Game status (active, cancelled, completed)
        type: string
      updated_at:
        description: Timestamp when the game was last updated
        type: string
      venue_address:
        type: string
      venue_id:
        description: Foreign key to venues table
        type: integer
      venue_name:
        type: string
      visibility:
        description: Visibility (public or private)
        type: string
    type: object
...
  main.BulkRemoveTokensRequest:
    properties:
      tokens:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - tokens
    type: object
...
  main.PruneStaleTokensRequest:
    properties:
      older_than:
        type: string
    required:
    - older_than
    type: object
...
  main.RemovePushTokenRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
...
  main.SavePushTokenRequest:
    properties:
      device_info:
        type: object
      token:
        type: string
    required:
    - token
    type: object
...
  main.createReviewPayload:
    properties:
      comment:
        maxLength: 500
        type: string
      rating:
        maximum: 5
        minimum: 1
        type: integer
    required:
    - comment
    - rating
    type: object
...
  venuereviews.Review:
    properties:
      avatar_url:
        type: string
      comment:
        type: string
      created_at:
        type: string
      id:
        type: integer
      rating:
        description: 1-5
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
      user_name:
        description: Joined fields
        type: string
      venue_id:
        type: integer
    type: object
...
info:
  contact:
    email: Ongchen10sherpa@gmail.com
    name: fullstacksherpa
    url: https://www.fullstacksherpa.tech/
  description: API for Khel, a complete sport application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Khel API
paths:
...
  /games/{gameID}/shortlist:
    delete:
      consumes:
      - application/json
      description: Allows authenticated users to remove a game from their shortlist.
      parameters:
      - description: Game ID
        in: path
        name: gameID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Game removed from shortlist
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Bad Request: Invalid game ID or unauthenticated request'
          schema: {}
        "500":
          description: 'Internal Server Error: Could not remove shortlist'
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Remove a game from shortlist
      tags:
      - Shortlist_Games
    post:
      consumes:
      - application/json
      description: Allows authenticated users to add a game to their shortlist.
      parameters:
      - description: Game ID
        in: path
        name: gameID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Game added to shortlist
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Bad Request: Invalid game ID or unauthenticated request'
          schema: {}
        "500":
          description: 'Internal Server Error: Could not add shortlist'
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Add a game to shortlist
      tags:
      - Shortlist_Games
...
  /games/shortlist:
    get:
      consumes:
      - application/json
      description: Returns a list of games that the authenticated user has shortlisted.
      produces:
      - application/json
      responses:
        "200":
          description: List of shortlisted games
          schema:
            items:
              items:
                $ref: '#/definitions/games.ShortlistedGameDetail'
              type: array
            type: array
        "400":
          description: 'Bad Request: Unauthenticated request'
          schema: {}
        "500":
          description: 'Internal Server Error: Could not retrieve shortlist'
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Retrieve shortlisted games for the authenticated user
      tags:
      - Shortlist_Games
...
  /push-tokens:
    delete:
      consumes:
      - application/json
      description: Deletes a specific push token for the current user
      parameters:
      - description: Token to remove
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.RemovePushTokenRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Remove a push notification token
      tags:
      - Notifications
...
  /users/push-tokens:
    post:
      consumes:
      - application/json
      description: Stores or updates a user's Expo push token along with optional
        device info
      parameters:
      - description: Push token data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.SavePushTokenRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Save or update a push notification token
      tags:
      - Notifications
  /users/push-tokens/bulk-remove:
    post:
      consumes:
      - application/json
      description: Deletes multiple push tokens (admin-only)
      parameters:
      - description: Tokens to remove
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.BulkRemoveTokensRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Bulk remove push notification tokens
      tags:
      - Notifications
  /users/push-tokens/prune:
    post:
      consumes:
      - application/json
      description: Removes push tokens not updated within the specified duration (admin-only)
      parameters:
      - description: Duration criteria
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.PruneStaleTokensRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Prune stale push tokens
      tags:
      - Notifications
...
  /venues/{venueID}/reviews:
    get:
      consumes:
      - application/json
      description: Retrieves all reviews for a specific venue along with the total
        count and average rating.
      parameters:
      - description: Venue ID
        in: path
        name: venueID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Reviews, total review count, and average rating
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'Bad Request: Invalid venue ID'
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Retrieve reviews for a venue
      tags:
      - Venue
    post:
      consumes:
      - application/json
      description: Creates a new review for a specific venue. The review includes
        a rating and comment.
      parameters:
      - description: Venue ID
        in: path
        name: venueID
        required: true
        type: integer
      - description: Review payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.createReviewPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Review created successfully
          schema:
            $ref: '#/definitions/venuereviews.Review'
        "400":
          description: 'Bad Request: Invalid input'
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Create a review for a venue
      tags:
      - Venue
  /venues/{venueID}/reviews/{reviewID}:
    delete:
      consumes:
      - application/json
      description: Deletes a review for a venue. This operation is allowed only if
        the requester is the review owner.
      parameters:
      - description: Venue ID
        in: path
        name: venueID
        required: true
        type: integer
      - description: Review ID
        in: path
        name: reviewID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Review deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 'Bad Request: Invalid review ID'
          schema: {}
        "404":
          description: 'Not Found: Review not found'
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Delete a venue review
      tags:
      - Venue
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"